+++ test/Skeletest/AssertionsSpec.hs
@@ -26,8 +26,7 @@ spec = do
         , "spec = it \"should fail\" $ 1 `shouldBe` (2 :: Int)"
         ]
 
-      (code, stdout, stderr) <- runTests runner []
-      code `shouldBe` ExitFailure 1
+      (stdout, stderr) <- expectFailure $ runTests runner []
       stderr `shouldBe` ""
       stdout `shouldSatisfy` P.matchesSnapshot
 
@@ -45,8 +44,7 @@ spec = do
         , "spec = it \"should fail\" $ 1 `shouldNotBe` (1 :: Int)"
         ]
 
-      (code, stdout, stderr) <- runTests runner []
-      code `shouldBe` ExitFailure 1
+      (stdout, stderr) <- expectFailure $ runTests runner []
       stderr `shouldBe` ""
       stdout `shouldSatisfy` P.matchesSnapshot
 
@@ -65,8 +63,7 @@ spec = do
         , "spec = it \"should fail\" $ (-1) `shouldSatisfy` P.gt (0 :: Int)"
         ]
 
-      (code, stdout, stderr) <- runTests runner []
-      code `shouldBe` ExitFailure 1
+      (stdout, stderr) <- expectFailure $ runTests runner []
       stderr `shouldBe` ""
       stdout `shouldSatisfy` P.matchesSnapshot
 
@@ -85,8 +82,7 @@ spec = do
         , "spec = it \"should fail\" $ 1 `shouldNotSatisfy` P.gt (0 :: Int)"
         ]
 
-      (code, stdout, stderr) <- runTests runner []
-      code `shouldBe` ExitFailure 1
+      (stdout, stderr) <- expectFailure $ runTests runner []
       stderr `shouldBe` ""
       stdout `shouldSatisfy` P.matchesSnapshot
 
@@ -103,8 +99,7 @@ spec = do
         , "    1 `shouldBe` (2 :: Int)"
         ]
 
-      (code, stdout, stderr) <- runTests runner []
-      code `shouldBe` ExitFailure 1
+      (stdout, stderr) <- expectFailure $ runTests runner []
       stderr `shouldBe` ""
       stdout `shouldSatisfy` P.matchesSnapshot
 
@@ -119,8 +114,7 @@ spec = do
         , "spec = it \"should fail\" $ failTest \"error message\""
         ]
 
-      (code, stdout, stderr) <- runTests runner []
-      code `shouldBe` ExitFailure 1
+      (stdout, stderr) <- expectFailure $ runTests runner []
       stderr `shouldBe` ""
       stdout `shouldSatisfy` P.matchesSnapshot
 
@@ -141,8 +135,7 @@ spec = do
       , "expectGT x actual = actual `shouldSatisfy` P.gt x"
       ]
 
-    (code, stdout, stderr) <- runTests runner []
-    code `shouldBe` ExitFailure 1
+    (stdout, stderr) <- expectFailure $ runTests runner []
     stderr `shouldBe` ""
     stdout `shouldSatisfy` P.matchesSnapshot
 
@@ -159,8 +152,7 @@ spec = do
       , "  x `shouldBe` True"
       ]
 
-    (code, stdout, stderr) <- runTests runner []
-    code `shouldBe` ExitFailure 1
+    (stdout, stderr) <- expectFailure $ runTests runner []
     stderr `shouldBe` ""
     stdout `shouldSatisfy` P.matchesSnapshot
 
@@ -177,8 +169,7 @@ spec = do
       , "  pure ()"
       ]
 
-    (code, stdout, stderr) <- runTests runner []
-    code `shouldBe` ExitFailure 1
+    (stdout, stderr) <- expectFailure $ runTests runner []
     stderr `shouldBe` ""
     sanitizeTraceback stdout `shouldSatisfy` P.matchesSnapshot
 
@@ -188,8 +179,12 @@ sanitizeTraceback :: String -> String
 #if __GLASGOW_HASKELL__ == 910
 sanitizeTraceback s =
   let (pre, post) = break (Text.pack "HasCallStack backtrace:" `Text.isInfixOf`) $ Text.lines $ Text.pack s
-      (_, post2) = break (Text.pack "╚" `Text.isPrefixOf`) $ drop 1 post
-   in Text.unpack . Text.unlines $ pre ++ post2
+      (_, post2) = break (Text.pack "╰" `Text.isPrefixOf`) $ drop 1 post
+      post2' =
+        case post2 of
+          [] -> []
+          l : ls -> Text.take 80 l : ls
+   in Text.unpack . Text.unlines $ pre ++ post2'
 #else
 sanitizeTraceback = id
 #endif