+++ test/Skeletest/TestUtils/Integration.hs
@@ -13,7 +13,9 @@ module Skeletest.TestUtils.Integration (
 
   -- * runTests
   runTests,
+  expectCode,
   expectSuccess,
+  expectFailure,
 
   -- * Re-exports
   ExitCode (..),
@@ -107,15 +109,25 @@ runTests FixtureTestRunner{..} args = do
       ]
   setCWD dir p = p{cwd = Just dir}
 
-  sanitize = Text.unpack . stripControlChars . Text.strip . Text.pack
+  sanitize = Text.unpack . stripOverwrites . stripControlChars . Text.strip . Text.pack
+  stripOverwrites s =
+    case Text.breakOn "\r" s of
+      (_, "") -> s
+      (pre, post) -> Text.dropWhileEnd (/= '\n') pre <> stripOverwrites (Text.drop 1 post)
   stripControlChars s =
     case Text.breakOn "\x1b" s of
       (_, "") -> s
       (pre, post) -> pre <> stripControlChars (Text.drop 1 . Text.dropWhile (/= 'm') $ post)
 
-expectSuccess :: (HasCallStack) => IO (ExitCode, String, String) -> IO (String, String)
-expectSuccess m = do
+expectCode :: (HasCallStack) => ExitCode -> IO (ExitCode, String, String) -> IO (String, String)
+expectCode expected m = do
   (code, stdout, stderr) <- m
   context (unlines ["===== stdout =====", stdout, "===== stderr =====", stderr]) $
-    code `shouldBe` ExitSuccess
+    code `shouldBe` expected
   pure (stdout, stderr)
+
+expectSuccess :: (HasCallStack) => IO (ExitCode, String, String) -> IO (String, String)
+expectSuccess = expectCode ExitSuccess
+
+expectFailure :: (HasCallStack) => IO (ExitCode, String, String) -> IO (String, String)
+expectFailure = expectCode $ ExitFailure 1