+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+API changes:
+* Re-export `BoxSpec` and `BoxContent` from `Skeletest.Plugin`
+
+Runtime changes:
+* Tweak formatting of test failures/errors
+
 ## v0.3.0
 
 GHC support: